// src/shared/api/generated/schemas/bet.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Logistics Auctions API
 * OpenAPI spec version: 0.0.1
 */

export interface Bet {
  uuid: string;
  auctionUuid: string;
  carrier: string;
  price: number;
  priceWithVat: number;
  place: number;
  isWinner: boolean;
  isCancelled: boolean;
  /** @nullable */
  cancelReason?: string | null;
  createdAt: string;
  isMyBet: boolean;
}